# Adding support for multiple values per dogstatsd datagram