--- a/pkg/dogstatsd/convert_bench_test.go
+++ b/pkg/dogstatsd/convert_bench_test.go
@@ -7,27 +7,50 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/metrics"
 )
 
-func buildRawSample(tagCount int) []byte {
+func buildRawSample(tagCount int, multipleValues bool) []byte {
 	tags := "tag0:val0"
 	for i := 1; i < tagCount; i++ {
 		tags += fmt.Sprintf(",tag%d:val%d", i, i)
 	}
 
+	if multipleValues {
+		return []byte(fmt.Sprintf("daemon:666:777|h|@0.5|#%s", tags))
+	}
 	return []byte(fmt.Sprintf("daemon:666|h|@0.5|#%s", tags))
 }
 
 // used to store the result and avoid optimizations
-var sample metrics.MetricSample
+var (
+	benchSamples []metrics.MetricSample
+)
 
-func BenchmarkParseMetric(b *testing.B) {
+func runParseMetricBenchmark(b *testing.B, multipleValues bool) {
+	parser := newParser(newFloat64ListPool())
+	namespaceBlacklist := []string{}
+
 	for i := 1; i < 1000; i *= 4 {
 		b.Run(fmt.Sprintf("%d-tags", i), func(sb *testing.B) {
-			rawSample := buildRawSample(i)
+			rawSample := buildRawSample(i, multipleValues)
 			sb.ResetTimer()
+			samples := make([]metrics.MetricSample, 0, 2)
 
 			for n := 0; n < sb.N; n++ {
-				sample, _ = parseAndEnrichMetricMessage(rawSample, "", []string{}, "default-hostname")
+
+				parsed, err := parser.parseMetricSample(rawSample)
+				if err != nil {
+					continue
+				}
+
+				benchSamples = enrichMetricSample(samples, parsed, "", namespaceBlacklist, "default-hostname", returnEmptyTags, true, false)
 			}
 		})
 	}
+}
+
+func BenchmarkParseMetric(b *testing.B) {
+	runParseMetricBenchmark(b, false)
+}
+
+func BenchmarkParseMultipleMetric(b *testing.B) {
+	runParseMetricBenchmark(b, true)
 }