--- a/pkg/dogstatsd/parse.go
+++ b/pkg/dogstatsd/parse.go
@@ -2,6 +2,7 @@ package dogstatsd
 
 import (
 	"bytes"
+	"fmt"
 	"strconv"
 	"unsafe"
 
@@ -28,14 +29,16 @@ var (
 // parser parses dogstatsd messages
 // not safe for concurent use
 type parser struct {
-	interner *stringInterner
+	interner    *stringInterner
+	float64List *float64ListPool
 }
 
-func newParser() *parser {
+func newParser(float64List *float64ListPool) *parser {
 	stringInternerCacheSize := config.Datadog.GetInt("dogstatsd_string_interner_size")
 
 	return &parser{
-		interner: newStringInterner(stringInternerCacheSize),
+		interner:    newStringInterner(stringInternerCacheSize),
+		float64List: float64List,
 	}
 }
 
@@ -80,6 +83,109 @@ func (p *parser) parseTags(rawTags []byte) []string {
 	}
 	tagsList[i] = p.interner.LoadOrStore(rawTags)
 	return tagsList
+}
+
+func (p *parser) parseMetricSample(message []byte) (dogstatsdMetricSample, error) {
+	// fast path to eliminate most of the gibberish
+	// especially important here since all the unidentified garbage gets
+	// identified as metrics
+	if !hasMetricSampleFormat(message) {
+		return dogstatsdMetricSample{}, fmt.Errorf("invalid dogstatsd message format")
+	}
+
+	rawNameAndValue, message := nextField(message)
+	name, rawValue, err := parseMetricSampleNameAndRawValue(rawNameAndValue)
+	if err != nil {
+		return dogstatsdMetricSample{}, err
+	}
+
+	rawMetricType, message := nextField(message)
+	metricType, err := parseMetricSampleMetricType(rawMetricType)
+	if err != nil {
+		return dogstatsdMetricSample{}, err
+	}
+
+	var setValue []byte
+	var values []float64
+	var value float64
+	if metricType == setType {
+		setValue = rawValue
+	} else {
+		// In case the list contains only one value, dogstatsd 1.0
+		// protocol, we directly parse it as a float64. This avoids
+		// pulling a slice from the float64List and greatly improve
+		// performances.
+		if bytes.Index(rawValue, colonSeparator) == -1 {
+			value, err = parseFloat64(rawValue)
+		} else {
+			values, err = p.parseFloat64List(rawValue)
+		}
+		if err != nil {
+			return dogstatsdMetricSample{}, fmt.Errorf("could not parse dogstatsd metric values: %v", err)
+		}
+	}
+
+	sampleRate := 1.0
+	var tags []string
+	var optionalField []byte
+	for message != nil {
+		optionalField, message = nextField(message)
+		if bytes.HasPrefix(optionalField, tagsFieldPrefix) {
+			tags = p.parseTags(optionalField[1:])
+		} else if bytes.HasPrefix(optionalField, sampleRateFieldPrefix) {
+			sampleRate, err = parseMetricSampleSampleRate(optionalField[1:])
+			if err != nil {
+				return dogstatsdMetricSample{}, fmt.Errorf("could not parse dogstatsd sample rate %q", optionalField)
+			}
+		}
+	}
+
+	return dogstatsdMetricSample{
+		name:       p.interner.LoadOrStore(name),
+		value:      value,
+		values:     values,
+		setValue:   string(setValue),
+		metricType: metricType,
+		sampleRate: sampleRate,
+		tags:       tags,
+	}, nil
+}
+
+// parseFloat64List parses a list of float64 separated by colonSeparator.
+func (p *parser) parseFloat64List(rawFloats []byte) ([]float64, error) {
+	var value float64
+	var err error
+	idx := 0
+
+	values := p.float64List.get()
+	for idx != -1 && len(rawFloats) != 0 {
+		idx = bytes.Index(rawFloats, colonSeparator)
+		// skip empty value such as '21::22'
+		if idx == 0 {
+			rawFloats = rawFloats[len(colonSeparator):]
+			continue
+		}
+
+		// last value
+		if idx == -1 {
+			value, err = parseFloat64(rawFloats)
+		} else {
+			value, err = parseFloat64(rawFloats[0:idx])
+			rawFloats = rawFloats[idx+len(colonSeparator):]
+		}
+
+		if err != nil {
+			p.float64List.put(values)
+			return nil, err
+		}
+
+		values = append(values, value)
+	}
+	if len(values) == 0 {
+		p.float64List.put(values)
+		return nil, fmt.Errorf("no value found")
+	}
+	return values, nil
 }
 
 // the std API does not have methods to do []byte => float parsing