--- a/pkg/dogstatsd/parse_events_test.go
+++ b/pkg/dogstatsd/parse_events_test.go
@@ -8,7 +8,7 @@ import (
 )
 
 func parseEvent(rawEvent []byte) (dogstatsdEvent, error) {
-	parser := newParser()
+	parser := newParser(newFloat64ListPool())
 	return parser.parseEvent(rawEvent)
 }