--- a/pkg/dogstatsd/parse_metrics.go
+++ b/pkg/dogstatsd/parse_metrics.go
@@ -29,8 +29,12 @@ var (
 )
 
 type dogstatsdMetricSample struct {
-	name       string
-	value      float64
+	name string
+	// use for single value messages
+	value float64
+	// use for multiple value messages
+	values []float64
+	// use to store set's values
 	setValue   string
 	metricType metricType
 	sampleRate float64
@@ -82,60 +86,4 @@ func parseMetricSampleMetricType(rawMetricType []byte) (metricType, error) {
 
 func parseMetricSampleSampleRate(rawSampleRate []byte) (float64, error) {
 	return parseFloat64(rawSampleRate)
-}
-
-func (p *parser) parseMetricSample(message []byte) (dogstatsdMetricSample, error) {
-	// fast path to eliminate most of the gibberish
-	// especially important here since all the unidentified garbage gets
-	// identified as metrics
-	if !hasMetricSampleFormat(message) {
-		return dogstatsdMetricSample{}, fmt.Errorf("invalid dogstatsd message format")
-	}
-
-	rawNameAndValue, message := nextField(message)
-	name, rawValue, err := parseMetricSampleNameAndRawValue(rawNameAndValue)
-	if err != nil {
-		return dogstatsdMetricSample{}, err
-	}
-
-	rawMetricType, message := nextField(message)
-	metricType, err := parseMetricSampleMetricType(rawMetricType)
-	if err != nil {
-		return dogstatsdMetricSample{}, err
-	}
-
-	var setValue []byte
-	var value float64
-	if metricType == setType {
-		setValue = rawValue
-	} else {
-		value, err = parseFloat64(rawValue)
-		if err != nil {
-			return dogstatsdMetricSample{}, fmt.Errorf("could not parse dogstatsd metric value: %v", err)
-		}
-	}
-
-	sampleRate := 1.0
-	var tags []string
-	var optionalField []byte
-	for message != nil {
-		optionalField, message = nextField(message)
-		if bytes.HasPrefix(optionalField, tagsFieldPrefix) {
-			tags = p.parseTags(optionalField[1:])
-		} else if bytes.HasPrefix(optionalField, sampleRateFieldPrefix) {
-			sampleRate, err = parseMetricSampleSampleRate(optionalField[1:])
-			if err != nil {
-				return dogstatsdMetricSample{}, fmt.Errorf("could not parse dogstatsd sample rate %q", optionalField)
-			}
-		}
-	}
-
-	return dogstatsdMetricSample{
-		name:       p.interner.LoadOrStore(name),
-		value:      value,
-		setValue:   string(setValue),
-		metricType: metricType,
-		sampleRate: sampleRate,
-		tags:       tags,
-	}, nil
 }