--- a/pkg/dogstatsd/parse_metrics_test.go
+++ b/pkg/dogstatsd/parse_metrics_test.go
@@ -8,7 +8,7 @@ import (
 )
 
 func parseMetricSample(rawSample []byte) (dogstatsdMetricSample, error) {
-	parser := newParser()
+	parser := newParser(newFloat64ListPool())
 	return parser.parseMetricSample(rawSample)
 }
 
@@ -20,9 +20,25 @@ func TestParseGauge(t *testing.T) {
 	assert.NoError(t, err)
 
 	assert.Equal(t, "daemon", sample.name)
+	assert.Equal(t, 666.0, sample.value)
 	assert.InEpsilon(t, 666.0, sample.value, epsilon)
+	require.Nil(t, sample.values)
 	assert.Equal(t, gaugeType, sample.metricType)
-	assert.Equal(t, 0, len(sample.tags))
+	assert.Len(t, sample.tags, 0)
+	assert.InEpsilon(t, 1.0, sample.sampleRate, epsilon)
+}
+
+func TestParseGaugeMultiple(t *testing.T) {
+	sample, err := parseMetricSample([]byte("daemon:666:777|g"))
+
+	assert.NoError(t, err)
+
+	assert.Equal(t, "daemon", sample.name)
+	assert.Len(t, sample.values, 2)
+	assert.InEpsilon(t, 666.0, sample.values[0], epsilon)
+	assert.InEpsilon(t, 777.0, sample.values[1], epsilon)
+	assert.Equal(t, gaugeType, sample.metricType)
+	assert.Len(t, sample.tags, 0)
 	assert.InEpsilon(t, 1.0, sample.sampleRate, epsilon)
 }
 
@@ -32,9 +48,24 @@ func TestParseCounter(t *testing.T) {
 	assert.NoError(t, err)
 
 	assert.Equal(t, "daemon", sample.name)
-	assert.Equal(t, 21.0, sample.value)
+	assert.InEpsilon(t, 21.0, sample.value, epsilon)
+	require.Nil(t, sample.values)
 	assert.Equal(t, countType, sample.metricType)
-	assert.Equal(t, 0, len(sample.tags))
+	assert.Len(t, sample.tags, 0)
+	assert.InEpsilon(t, 1.0, sample.sampleRate, epsilon)
+}
+
+func TestParseCounterMultiple(t *testing.T) {
+	sample, err := parseMetricSample([]byte("daemon:666:777|c"))
+
+	assert.NoError(t, err)
+
+	assert.Equal(t, "daemon", sample.name)
+	assert.Len(t, sample.values, 2)
+	assert.InEpsilon(t, 666.0, sample.values[0], epsilon)
+	assert.InEpsilon(t, 777.0, sample.values[1], epsilon)
+	assert.Equal(t, countType, sample.metricType)
+	assert.Len(t, sample.tags, 0)
 	assert.InEpsilon(t, 1.0, sample.sampleRate, epsilon)
 }
 
@@ -44,7 +75,8 @@ func TestParseCounterWithTags(t *testing.T) {
 	assert.NoError(t, err)
 
 	assert.Equal(t, "custom_counter", sample.name)
-	assert.Equal(t, 1.0, sample.value)
+	assert.InEpsilon(t, 1.0, sample.value, epsilon)
+	require.Nil(t, sample.values)
 	assert.Equal(t, countType, sample.metricType)
 	assert.Equal(t, 2, len(sample.tags))
 	assert.Equal(t, "protocol:http", sample.tags[0])
@@ -58,9 +90,24 @@ func TestParseHistogram(t *testing.T) {
 	assert.NoError(t, err)
 
 	assert.Equal(t, "daemon", sample.name)
-	assert.Equal(t, 21.0, sample.value)
+	assert.InEpsilon(t, 21.0, sample.value, epsilon)
+	require.Nil(t, sample.values)
 	assert.Equal(t, histogramType, sample.metricType)
-	assert.Equal(t, 0, len(sample.tags))
+	assert.Len(t, sample.tags, 0)
+	assert.InEpsilon(t, 1.0, sample.sampleRate, epsilon)
+}
+
+func TestParseHistogramrMultiple(t *testing.T) {
+	sample, err := parseMetricSample([]byte("daemon:21:22|h"))
+
+	assert.NoError(t, err)
+
+	assert.Equal(t, "daemon", sample.name)
+	assert.Len(t, sample.values, 2)
+	assert.InEpsilon(t, 21.0, sample.values[0], epsilon)
+	assert.InEpsilon(t, 22.0, sample.values[1], epsilon)
+	assert.Equal(t, histogramType, sample.metricType)
+	assert.Len(t, sample.tags, 0)
 	assert.InEpsilon(t, 1.0, sample.sampleRate, epsilon)
 }
 
@@ -70,9 +117,24 @@ func TestParseTimer(t *testing.T) {
 	assert.NoError(t, err)
 
 	assert.Equal(t, "daemon", sample.name)
-	assert.Equal(t, 21.0, sample.value)
+	assert.InEpsilon(t, 21.0, sample.value, epsilon)
+	require.Nil(t, sample.values)
 	assert.Equal(t, timingType, sample.metricType)
-	assert.Equal(t, 0, len(sample.tags))
+	assert.Len(t, sample.tags, 0)
+	assert.InEpsilon(t, 1.0, sample.sampleRate, epsilon)
+}
+
+func TestParseTimerMultiple(t *testing.T) {
+	sample, err := parseMetricSample([]byte("daemon:21:22|ms"))
+
+	assert.NoError(t, err)
+
+	assert.Equal(t, "daemon", sample.name)
+	assert.Len(t, sample.values, 2)
+	assert.InEpsilon(t, 21.0, sample.values[0], epsilon)
+	assert.InEpsilon(t, 22.0, sample.values[1], epsilon)
+	assert.Equal(t, timingType, sample.metricType)
+	assert.Len(t, sample.tags, 0)
 	assert.InEpsilon(t, 1.0, sample.sampleRate, epsilon)
 }
 
@@ -84,19 +146,47 @@ func TestParseSet(t *testing.T) {
 	assert.Equal(t, "daemon", sample.name)
 	assert.Equal(t, "abc", sample.setValue)
 	assert.Equal(t, setType, sample.metricType)
-	assert.Equal(t, 0, len(sample.tags))
+	assert.Len(t, sample.tags, 0)
 	assert.InEpsilon(t, 1.0, sample.sampleRate, epsilon)
 }
 
-func TestSampleistribution(t *testing.T) {
+func TestParseSetMultiple(t *testing.T) {
+	// multiple values are not supported for set. ':' can be part of the
+	// set value for backward compatibility
+	sample, err := parseMetricSample([]byte("daemon:abc:def|s"))
+
+	assert.NoError(t, err)
+
+	assert.Equal(t, "daemon", sample.name)
+	assert.Equal(t, "abc:def", sample.setValue)
+	assert.Equal(t, setType, sample.metricType)
+	assert.Len(t, sample.tags, 0)
+	assert.InEpsilon(t, 1.0, sample.sampleRate, epsilon)
+}
+
+func TestSampleDistribution(t *testing.T) {
 	sample, err := parseMetricSample([]byte("daemon:3.5|d"))
 
 	assert.NoError(t, err)
 
 	assert.Equal(t, "daemon", sample.name)
-	assert.Equal(t, 3.5, sample.value)
+	assert.InEpsilon(t, 3.5, sample.value, epsilon)
+	require.Nil(t, sample.values)
 	assert.Equal(t, distributionType, sample.metricType)
-	assert.Equal(t, 0, len(sample.tags))
+	assert.Len(t, sample.tags, 0)
+}
+
+func TestParseDistributionMultiple(t *testing.T) {
+	sample, err := parseMetricSample([]byte("daemon:3.5:4.5|d"))
+
+	assert.NoError(t, err)
+
+	assert.Equal(t, "daemon", sample.name)
+	assert.Len(t, sample.values, 2)
+	assert.InEpsilon(t, 3.5, sample.values[0], epsilon)
+	assert.InEpsilon(t, 4.5, sample.values[1], epsilon)
+	assert.Equal(t, distributionType, sample.metricType)
+	assert.Len(t, sample.tags, 0)
 }
 
 func TestParseSetUnicode(t *testing.T) {
@@ -107,7 +197,7 @@ func TestParseSetUnicode(t *testing.T) {
 	assert.Equal(t, "daemon", sample.name)
 	assert.Equal(t, "♬†øU†øU¥ºuT0♪", sample.setValue)
 	assert.Equal(t, setType, sample.metricType)
-	assert.Equal(t, 0, len(sample.tags))
+	assert.Len(t, sample.tags, 0)
 	assert.InEpsilon(t, 1.0, sample.sampleRate, epsilon)
 }
 
@@ -118,6 +208,7 @@ func TestParseGaugeWithTags(t *testing.T) {
 
 	assert.Equal(t, "daemon", sample.name)
 	assert.InEpsilon(t, 666.0, sample.value, epsilon)
+	require.Nil(t, sample.values)
 	assert.Equal(t, gaugeType, sample.metricType)
 	require.Equal(t, 2, len(sample.tags))
 	assert.Equal(t, "sometag1:somevalue1", sample.tags[0])
@@ -131,6 +222,7 @@ func TestParseGaugeWithNoTags(t *testing.T) {
 
 	assert.Equal(t, "daemon", sample.name)
 	assert.InEpsilon(t, 666.0, sample.value, epsilon)
+	require.Nil(t, sample.values)
 	assert.Equal(t, gaugeType, sample.metricType)
 	assert.Empty(t, sample.tags)
 	assert.InEpsilon(t, 1.0, sample.sampleRate, epsilon)
@@ -143,8 +235,9 @@ func TestParseGaugeWithSampleRate(t *testing.T) {
 
 	assert.Equal(t, "daemon", sample.name)
 	assert.InEpsilon(t, 666.0, sample.value, epsilon)
+	require.Nil(t, sample.values)
 	assert.Equal(t, gaugeType, sample.metricType)
-	assert.Equal(t, 0, len(sample.tags))
+	assert.Len(t, sample.tags, 0)
 	assert.InEpsilon(t, 0.21, sample.sampleRate, epsilon)
 }
 
@@ -155,8 +248,9 @@ func TestParseGaugeWithPoundOnly(t *testing.T) {
 
 	assert.Equal(t, "daemon", sample.name)
 	assert.InEpsilon(t, 666.0, sample.value, epsilon)
+	require.Nil(t, sample.values)
 	assert.Equal(t, gaugeType, sample.metricType)
-	assert.Equal(t, 0, len(sample.tags))
+	assert.Len(t, sample.tags, 0)
 	assert.InEpsilon(t, 1.0, sample.sampleRate, epsilon)
 }
 
@@ -167,6 +261,7 @@ func TestParseGaugeWithUnicode(t *testing.T) {
 
 	assert.Equal(t, "♬†øU†øU¥ºuT0♪", sample.name)
 	assert.InEpsilon(t, 666.0, sample.value, epsilon)
+	require.Nil(t, sample.values)
 	assert.Equal(t, gaugeType, sample.metricType)
 	require.Equal(t, 1, len(sample.tags))
 	assert.Equal(t, "intitulé:T0µ", sample.tags[0])
@@ -188,10 +283,6 @@ func TestParseMetricError(t *testing.T) {
 	assert.Error(t, err)
 
 	_, err = parseMetricSample([]byte("abc666|g"))
-	assert.Error(t, err)
-
-	// too many value
-	_, err = parseMetricSample([]byte("daemon:666:777|g"))
 	assert.Error(t, err)
 
 	// unknown metadata prefix