--- a/pkg/dogstatsd/parse_service_checks_test.go
+++ b/pkg/dogstatsd/parse_service_checks_test.go
@@ -8,7 +8,7 @@ import (
 )
 
 func parseServiceCheck(rawServiceCheck []byte) (dogstatsdServiceCheck, error) {
-	parser := newParser()
+	parser := newParser(newFloat64ListPool())
 	return parser.parseServiceCheck(rawServiceCheck)
 }