### What does this PR do?

Numerical metrics (ie: not 'set') in dogstatsd now support sending
multiple values per message. This allows client to reduce traffic usage
and enable client side aggregation for histogram and distribution.

### Performances

I tried to keep the same performance for single value messages than before. For this we allocate slice of samples as early as possible and pass them around. 

To compare to master I had to rewrite a bit some benchmark to take the GC out of it. This [PR](https://github.com/DataDog/datadog-agent/pull/6832/files) has been use as the reference for master.

Benchmarks where run with a 5s benchtime (`go test -bench=. -test.run=nothing -benchtime=5s`) on a `t2.xlarge` (4 vCPUs and 16GB or ram) on a Ubuntu with `go1.15.5`.

We have two tool to compare benchmark


#### benchstat

```
name                    old time/op    new time/op    delta
ParseMetric/1-tags-4       447ns ± 0%     469ns ± 0%   ~     (p=1.000 n=1+1)
ParseMetric/4-tags-4       694ns ± 0%     699ns ± 0%   ~     (p=1.000 n=1+1)
ParseMetric/16-tags-4     1.60µs ± 0%    1.54µs ± 0%   ~     (p=1.000 n=1+1)
ParseMetric/64-tags-4     5.06µs ± 0%    5.14µs ± 0%   ~     (p=1.000 n=1+1)
ParseMetric/256-tags-4    21.0µs ± 0%    21.7µs ± 0%   ~     (p=1.000 n=1+1)
EnrichTags/20-tags-4       229ns ± 0%     231ns ± 0%   ~     (p=1.000 n=1+1)
EnrichTags/40-tags-4       317ns ± 0%     320ns ± 0%   ~     (p=1.000 n=1+1)
EnrichTags/60-tags-4       412ns ± 0%     419ns ± 0%   ~     (p=1.000 n=1+1)
EnrichTags/80-tags-4       518ns ± 0%     522ns ± 0%   ~     (p=1.000 n=1+1)
EnrichTags/100-tags-4      580ns ± 0%     585ns ± 0%   ~     (p=1.000 n=1+1)
EnrichTags/120-tags-4      701ns ± 0%     710ns ± 0%   ~     (p=1.000 n=1+1)
EnrichTags/140-tags-4      780ns ± 0%     867ns ± 0%   ~     (p=1.000 n=1+1)
EnrichTags/160-tags-4      902ns ± 0%     976ns ± 0%   ~     (p=1.000 n=1+1)
EnrichTags/180-tags-4     1.04µs ± 0%    0.97µs ± 0%   ~     (p=1.000 n=1+1)
EnrichTags/200-tags-4     1.08µs ± 0%    1.10µs ± 0%   ~     (p=1.000 n=1+1)
ParsePackets-4             238µs ± 0%     256µs ± 0%   ~     (p=1.000 n=1+1)
ParseMetricMessage-4       278ns ± 0%     257ns ± 0%   ~     (p=1.000 n=1+1)
WithMapper-4              4.03µs ± 0%    4.27µs ± 0%   ~     (p=1.000 n=1+1)
MapperControl-4           4.05µs ± 0%    4.31µs ± 0%   ~     (p=1.000 n=1+1)

name                    old alloc/op   new alloc/op   delta
WithMapper-4              4.59kB ± 0%    4.59kB ± 0%   ~     (all equal)
MapperControl-4           4.59kB ± 0%    4.59kB ± 0%   ~     (all equal)

name                    old allocs/op  new allocs/op  delta
WithMapper-4                14.0 ± 0%      14.0 ± 0%   ~     (all equal)
MapperControl-4             14.0 ± 0%      14.0 ± 0%   ~     (all equal)
```

#### benchcmp (deprecated)

`benchcmp` should be disregarded in favor or `benchstat` as running the benchmark on master twice can produce up to 10% difference for some bench (especially for smaller bench running around ~500ns/op).

```
benchcmp is deprecated in favor of benchstat: https://pkg.go.dev/golang.org/x/perf/cmd/benchstat
benchmark                           old ns/op     new ns/op     delta
BenchmarkParseMetric/1-tags-4       447           469           +4.92%
BenchmarkParseMetric/4-tags-4       694           699           +0.72%
BenchmarkParseMetric/16-tags-4      1599          1536          -3.94%
BenchmarkParseMetric/64-tags-4      5059          5144          +1.68%
BenchmarkParseMetric/256-tags-4     21031         21717         +3.26%
BenchmarkEnrichTags/20-tags-4       229           231           +0.87%
BenchmarkEnrichTags/40-tags-4       317           320           +0.95%
BenchmarkEnrichTags/60-tags-4       412           419           +1.70%
BenchmarkEnrichTags/80-tags-4       518           522           +0.77%
BenchmarkEnrichTags/100-tags-4      580           585           +0.86%
BenchmarkEnrichTags/120-tags-4      701           710           +1.28%
BenchmarkEnrichTags/140-tags-4      780           867           +11.15%
BenchmarkEnrichTags/160-tags-4      902           976           +8.20%
BenchmarkEnrichTags/180-tags-4      1041          967           -7.11%
BenchmarkEnrichTags/200-tags-4      1077          1095          +1.67%
BenchmarkParsePackets-4             237536        256212        +7.86%
BenchmarkParseMetricMessage-4       278           257           -7.55%
BenchmarkWithMapper-4               4029          4272          +6.03%
BenchmarkMapperControl-4            4048          4311          +6.50%

benchmark                    old allocs     new allocs     delta
BenchmarkWithMapper-4        14             14             +0.00%
BenchmarkMapperControl-4     14             14             +0.00%

benchmark                    old bytes     new bytes     delta
BenchmarkWithMapper-4        4588          4588          +0.00%
BenchmarkMapperControl-4     4588          4588          +0.00%
```