@verdie-g 2020-11-16
Q: Using this feature in a DogStatsD client when the agent version doesn't support it will result in discarded metrics. How do you think DogStatsD clients could safely expose this new feature?