@hush-hush 2020-11-16
@verdie-g great question. At first this option will be disable by default on the client side leaving it up to the users enable it when they can.

For later, we're looking into ways for the client to fetch the agent version. But this will come after full client side aggregation is done for at least for Go, Java and C# client.