@verdie-g 2020-11-16
> For later, we're looking into ways for the client to fetch the agent version.

That's interesting but won't it add much complexity in both client and server? The communication between the client and the agent will become bidirectional when it was only unidirectional. (the server only reads from a socket and never writes). Also the clients will have to handle dropped UDP frames, unavailable agent when requesting the version.

> At first this option will be disable by default on the client side leaving it up to the users enable it when they can.

Rather than a flag, for [my library](https://github.com/verdie-g/DatadogStatsD) I think I'll ask the user to specify the targeted agent version and enable this feature (and maybe future ones) depending on that.