@hush-hush 2020-11-16
> That's interesting but won't it add much complexity in both client and server? The communication between the client and the agent will become bidirectional when it was only unidirectional. (the server only reads from a socket and never writes). Also the clients will have to handle dropped UDP frames, unavailable agent when requesting the version.

Yes you're right. For now we haven't chose a precise solution yet to sync both agent and clients. We started to add aggregation for `count`, `gauge` and `set` (in beta in Go client). Multiple values per message is the next step we are exploring for `histogram` and `distribution`.

> Rather than a flag, for [my library](https://github.com/verdie-g/DatadogStatsD) I think I'll ask the user to specify the targeted agent version and enable this feature (and maybe future ones) depending on that.

 Specifying a DogStastD protocol version in the client's option could also be a option too.