@verdie-g 2020-11-17
Second question: for telemetry, is a multi-values datagram considered as one metric?