@hush-hush 2020-11-19
> Second question: for telemetry, is a multi-value datagram considered as one metric?

No we have the difference between packets and metrics. A metrics is a sample call from the client and a packet is what is sent on the wire.

So for example 10 call to `Increment()` will produce `datadog.dogstatsd.client.metrics` to 10 but `datadog.dogstatsd.client.packets_sent` to one (as the 10 counts produce by `Increment()` will fit in one packet).

You can look at the Go implementation if you want more details: https://github.com/DataDog/datadog-go