@hush-hush 2020-12-09
@verdie-g this has been merged. Just a quick highlight to remind you it's still in an experimental state in case you want to use it in your repo.

The plan is to move it as stable in a later release of the agent.